implemented use cases for retriving fibonacci numbers

+++ src/Entity/DTO/FibonacciSubsequence/FibonacciSubsequenceResponse.cs
@@ -2,5 +2,5 @@ namespace Entity.DTO.FibonacciSubsequence;
 
 public sealed class FibonacciSubsequenceResponse
 {
-    public int[] Subsequence { get; set; }
+    public long[] Subsequence { get; set; }
 }
